remove unused ci types
