+++ projects/openapi-utilities/src/legacy-ci-types.ts
@@ -1,0 +1,10 @@
+import { ResultWithSourcemap } from './types';
+import { IChange } from './openapi3/sdk/types';
+
+export type CompareFileJson = {
+  results: ResultWithSourcemap[];
+  changes: IChange[];
+  // TODO remove this on 2022 Aug
+  projectRootDir?: string | false;
+  version?: string;
+};
